requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
yfinance
requests
python-dotenv
langchain>=0.2.14
langchain-openai>=0.1.23
langchain-core>=0.2.35
langgraph>=0.2.14
langchain-community
chromadb
sentence-transformers
newsapi-python
pdfplumber
jupyter
tiktoken
qdrant-client
arxiv
langchain-text-splitters
lxml